refactor lexer to store pointer to input
refactored lexer class to get rid of input string as an argument to all
methods except init. Init now stores a pointer to input string as part
of class data. This cleans up the function call chain a lot and we now
no longer have to pass the input string as a an argument to every dang
function